hw/i2c/aspeed_i2c: Make AspeedI2CClass::gap an plain unsigned type
Convert AspeedI2CClass::gap to plain unsigned, using '0'
as "no gap" to avoid the followin UBSan warnings:

  SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior ../../hw/i2c/aspeed_i2c.c:1559:16
  SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior ../../hw/i2c/aspeed_i2c.c:1583:16
  SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior ../../hw/i2c/aspeed_i2c.c:1608:16
  hw/i2c/aspeed_i2c.c:1608:16: runtime error: implicit conversion from type 'int' of value
                               -1 (32-bit, signed) to type 'uint8_t' (aka 'unsigned char')
                               changed the value to 255 (8-bit, unsigned)

Message-Id: <20250423120555.21318-1-philmd@linaro.org>

+++ include/hw/i2c/aspeed_i2c.h
@@ -290,7 +290,7 @@ struct AspeedI2CClass {
     uint8_t num_busses;
     uint8_t reg_size;
     uint32_t reg_gap_size;
-    uint8_t gap;
+    unsigned gap;
     qemu_irq (*bus_get_irq)(AspeedI2CBus *);
 
     uint64_t pool_size;
